Redirect to frontend main page

+++ src/main/java/com/meetme/meetme/service/EventServiceImpl.java
@@ -4,6 +4,7 @@ import com.meetme.meetme.mapper.EventMapper;
 import com.meetme.meetme.model.EventDto;
 import com.meetme.meetme.repository.EventRepository;
 import lombok.AllArgsConstructor;
+import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -28,30 +29,47 @@ public class EventServiceImpl implements EventService {
     @Override
     public List<JSONObject> getEvents() {
         List<EventDto> events =  eventRepository.findAll().stream().map(eventMapper::eventToEventDto).toList();
+
         List<JSONObject> eventsList = new ArrayList<>();
 
         int i = 0;
         for (EventDto e:events){
             JSONObject event = new JSONObject();
 
-            event.put("title", e.getTitle());
+            if (e.getTitle() != null){
+                event.put("title", e.getTitle());
+            }
 
-            JSONObject start_date = new JSONObject();
-            start_date.put("start_date", e.getDate());
-            event.put("date", start_date);
+            if (e.getDate() != null){
+                JSONObject start_date = new JSONObject();
+                start_date.put("start_date", e.getDate());
+                event.put("date", start_date);
+            }
 
-            String[] address = new String[]{e.getLocation()};
-            event.put("address", address);
+            if (e.getLocation() != null){
+                String[] address = new String[]{e.getLocation()};
+                event.put("address", address);
+            }
 
-            event.put("link", e.getLink());
+            if (e.getLink() != null){
+                event.put("link", e.getLink());
+            }
 
-            event.put("ticket_info", e.getTicketInfo());
+            if (e.getTicketInfo() != null){
+                JSONArray ticketInfo = new JSONArray(e.getTicketInfo());
+                event.put("ticket_info", ticketInfo);
+            }
 
-            event.put("description", e.getDescription());
+            if (e.getDescription() != null){
+                event.put("description", e.getDescription());
+            }
 
-            event.put("image", e.getPhotoLink());
+            if (e.getPhotoLink() != null){
+                event.put("image", e.getPhotoLink());
+                event.put("thumbnail", e.getPhotoLink());
+            }
 
-            String coordinates = "["+e.getLng()+", "+e.getLng()+"]";
+            String coordinates = "["+e.getLng()+", "+e.getLat()+"]";
             event.put("coordinates", coordinates);
 
             event.put("id", i);
@@ -59,13 +77,11 @@ public class EventServiceImpl implements EventService {
 
             eventsList.add(event);
         }
-
         return eventsList;
     }
 
     @Override
     public EventDto saveNewEvent(EventDto eventDto) {
-
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         eventDto.setUser(userService.getByUsername(authentication.getName()));
         return eventMapper.eventToEventDto(eventRepository.save(eventMapper.eventDtoToEvent(eventDto)));
@@ -73,9 +89,13 @@ public class EventServiceImpl implements EventService {
 
     @Override
     public Boolean deleteById(Long id) {
-        if(eventRepository.existsById(id)){
-            eventRepository.deleteById(id);
-            return true;
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if (eventRepository.existsById(id)) {
+            if (userService.getByUsername(authentication.getName()).getId().equals(eventRepository.findById(id).orElseThrow().getId())) {
+                eventRepository.deleteById(id);
+                return true;
+            }
         }
         return false;
     }
